# Optimized distinct() operator